level8: done - rework source.c

+++ level7/source.c
@@ -7,7 +7,7 @@ char c[68];
 
 void m()
 {
-    printf("%s - %d\n", c, time(0));
+    printf("%s - %d\n", c, (int)time(0));
 }
 
 // int main(int argc, char **argv)
@@ -36,11 +36,11 @@ int main(int argc, char **argv)
 
     ptr1 = malloc(8);
     ptr1[0] = 1;
-    ptr1[1] = malloc(8);
+    ptr1[1] = (unsigned int)malloc(8);
 
     ptr2 = malloc(8);
     ptr2[0] = 2;
-    ptr2[1] = malloc(8);
+    ptr2[1] = (unsigned int)malloc(8);
 
     strcpy((char *)ptr1[1], argv[1]);
     strcpy((char *)ptr2[1], argv[2]);
